migrate handleevent and reducer in the page without affecting memberlist component in gitness
