+++ packages/playground/src/components/project-settings/members-list.tsx
@@ -1,4 +1,4 @@
-import React, { useReducer } from 'react'
+import React from 'react'
 import {
   Icon,
   Table,
@@ -22,9 +22,6 @@ import {
   DropdownMenuTrigger
 } from '@harnessio/canary'
 import { getInitials } from '../../utils/utils'
-import { FormEditMemberDialog } from './form-member-edit-dialog'
-import { FormDeleteMemberDialog } from './form-member-delete-dialog'
-import { dialogStateReducer, initialDialogState } from './members-reducers/dialog-state-reducers'
 
 interface MembersProps {
   display_name: string
@@ -36,44 +33,11 @@ interface MembersProps {
 
 interface PageProps {
   members: MembersProps[]
+  onDelete: (member: MembersProps) => void
+  onEdit: (member: MembersProps) => void
 }
 
-export const MembersList = ({ members }: PageProps) => {
-  const [dialogState, dispatch] = useReducer(dialogStateReducer, initialDialogState)
-
-  const openDialog = (dialogType: 'edit' | 'delete', member: MembersProps) => {
-    dispatch({ type: 'OPEN_DIALOG', dialogType, member })
-  }
-
-  const closeDialog = (dialogType: 'edit' | 'delete') => {
-    dispatch({ type: 'CLOSE_DIALOG', dialogType })
-  }
-
-  // Delete project handler
-  const handleDelete = () => {
-    dispatch({ type: 'START_DELETING' })
-
-    setTimeout(() => {
-      dispatch({ type: 'DELETE_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('delete')
-        dispatch({ type: 'RESET_DELETE' })
-      }, 2000)
-    }, 2000)
-  }
-
-  const handleRoleSave = () => {
-    dispatch({ type: 'START_SUBMITTING' })
-
-    setTimeout(() => {
-      dispatch({ type: 'SUBMIT_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('edit')
-        dispatch({ type: 'RESET_SUBMIT' })
-      }, 2000)
-    }, 2000)
-  }
-
+export const MembersList = ({ members, onDelete, onEdit }: PageProps) => {
   const moreActionsTooltip = ({ member }: { member: MembersProps }) => {
     return (
       <DropdownMenu>
@@ -84,7 +48,7 @@ export const MembersList = ({ members }: PageProps) => {
         </DropdownMenuTrigger>
         <DropdownMenuContent className="shadow-sm py-2 bg-primary-background border border-gray-800 rounded-[10px] w-[180px]">
           <DropdownMenuGroup>
-            <DropdownMenuItem className="cursor-pointer" onSelect={() => openDialog('edit', member)}>
+            <DropdownMenuItem className="cursor-pointer" onSelect={() => onEdit(member)}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="edit-pen" className="mr-2" />
               </DropdownMenuShortcut>
@@ -93,7 +57,7 @@ export const MembersList = ({ members }: PageProps) => {
             <DropdownMenuSeparator />
             <DropdownMenuItem
               className="cursor-pointer text-red-400 hover:text-red-400 focus:text-red-400"
-              onSelect={() => openDialog('delete', member)}>
+              onSelect={() => onDelete(member)}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="trash" className="mr-2 text-red-400" />
               </DropdownMenuShortcut>
@@ -113,7 +77,7 @@ export const MembersList = ({ members }: PageProps) => {
             <TableHead className="text-primary">Name</TableHead>
             <TableHead className="text-primary">Role</TableHead>
             <TableHead className="text-primary">Email</TableHead>
-            {members[0]?.timestamp && <TableHead className="text-right text-primary">Date added</TableHead>}
+            {<TableHead className="text-right text-primary">Date added</TableHead>}
             <TableHead>
               <></>
             </TableHead>
@@ -174,32 +138,6 @@ export const MembersList = ({ members }: PageProps) => {
           ))}
         </TableBody>
       </Table>
-      {/* Delete Dialog */}
-      {dialogState.isDialogDeleteOpen && (
-        <FormDeleteMemberDialog
-          isDeleting={dialogState.isDeleting}
-          deleteSuccess={dialogState.deleteSuccess}
-          onDelete={handleDelete}
-          member={dialogState.deleteMember!}
-          onClose={() => {
-            closeDialog('delete')
-            dispatch({ type: 'RESET_DELETE' })
-          }}
-        />
-      )}
-      {/* Edit Dialog */}
-      {dialogState.isDialogEditOpen && dialogState.editMember && (
-        <FormEditMemberDialog
-          member={dialogState.editMember}
-          onSave={handleRoleSave}
-          onClose={() => {
-            closeDialog('edit')
-            dispatch({ type: 'RESET_SUBMIT' })
-          }}
-          isSubmitting={dialogState.isSubmitting}
-          submitted={dialogState.submitted}
-        />
-      )}
     </>
   )
 }
